🎨 remove duplicated code
