+++ src/functions/searches/helpers/attach-search.ts
@@ -1,4 +1,4 @@
-import { filter, firstValueFrom, Observable } from 'rxjs';
+import { filter, firstValueFrom, map, Observable } from 'rxjs';
 import { DateRange } from '~/functions';
 import { SearchFilter } from '~/main';
 import { filterMessageByCommand, RequiredSearchFilter } from '../../searches/subscribe-to-one-search/helpers';
@@ -7,6 +7,14 @@ import { SearchMessageCommands } from '~/models';
 import { RawSearchMessageReceived } from '~/models';
 import { APISubscription } from '~/functions/utils';
 import { RawSearchMessageSent } from '~/models';
+import { RawResponseForSearchStatsWithinRangeMessageReceived } from '~/models';
+import { countEntriesFromModules } from '../../searches/subscribe-to-one-search/helpers';
+import { isAfter } from 'date-fns';
+import { SearchFrequencyStats, SearchStats } from '~/models';
+import { RawResponseForSearchStatsMessageReceived, RawResponseForSearchDetailsMessageReceived } from '~/models';
+import { isUndefined } from 'lodash';
+import { toNumericID } from '../../../value-objects/id';
+import { RawSearchAttachedMessageReceived } from '~/models';
 
 // Dynamic duration for debounce a after each event, starting from 1s and increasing 500ms after each event,
 // never surpass 4s, reset to 1s if the request is finished
@@ -64,3 +72,118 @@ export const getPreviewDateRange: (props: {
 		end: new Date(details.data.SearchInfo.EndRange),
 	};
 };
+
+export const extractZoomFromRawSearchStats: (
+	props: Readonly<{
+		filtersByID: Record<string, SearchFilter | undefined>;
+		initialFilter: RequiredSearchFilter;
+		previewDateRange: DateRange;
+	}>,
+) => (
+	source: Observable<RawResponseForSearchStatsWithinRangeMessageReceived>,
+) => Observable<{
+	frequencyStats: Array<SearchFrequencyStats>;
+	filter: SearchFilter | undefined;
+}> = ({ filtersByID, initialFilter, previewDateRange }) => source =>
+	source.pipe(
+		map(set => {
+			const filterID = (set.data.Addendum?.filterID as string | undefined) ?? null;
+			const filter = filtersByID[filterID ?? ''] ?? undefined;
+
+			const filterEnd = filter?.dateRange === 'preview' ? previewDateRange.end : filter?.dateRange?.end;
+			const initialEnd = initialFilter.dateRange === 'preview' ? previewDateRange.end : initialFilter.dateRange.end;
+			const endDate = filterEnd ?? initialEnd;
+
+			return {
+				frequencyStats: countEntriesFromModules(set).filter(f => !isAfter(f.timestamp, endDate)),
+				filter,
+			};
+		}),
+	);
+
+export const mapToSearchStats: (
+	props: Readonly<{
+		filtersByID: Record<string, SearchFilter | undefined>;
+		searchAttachMsg: RawSearchAttachedMessageReceived;
+	}>,
+) => (
+	source: Observable<[RawResponseForSearchStatsMessageReceived, RawResponseForSearchDetailsMessageReceived]>,
+) => Observable<SearchStats> = ({ filtersByID, searchAttachMsg }) => source =>
+	source.pipe(
+		map(
+			([rawStats, rawDetails]): SearchStats => {
+				const filterID =
+					(rawStats.data.Addendum?.filterID as string | undefined) ??
+					(rawDetails.data.Addendum?.filterID as string | undefined) ??
+					null;
+				const filter = filtersByID[filterID ?? ''] ?? undefined;
+
+				const pipeline = rawStats.data.Stats.Set.map(s => s.Stats)
+					.reduce<
+						Array<Array<RawResponseForSearchStatsMessageReceived['data']['Stats']['Set'][number]['Stats'][number]>>
+					>((acc, curr) => {
+						curr.forEach((_curr, i) => {
+							if (isUndefined(acc[i])) acc[i] = [];
+							acc[i].push(_curr);
+						});
+						return acc;
+					}, [])
+					.map(s =>
+						s
+							.map(_s => ({
+								module: _s.Name,
+								arguments: _s.Args,
+								duration: _s.Duration,
+								input: {
+									bytes: _s.InputBytes,
+									entries: _s.InputCount,
+								},
+								output: {
+									bytes: _s.OutputBytes,
+									entries: _s.OutputCount,
+								},
+							}))
+							.reduce((acc, curr) => ({
+								...curr,
+								duration: acc.duration + curr.duration,
+								input: {
+									bytes: acc.input.bytes + curr.input.bytes,
+									entries: acc.input.entries + curr.input.entries,
+								},
+								output: {
+									bytes: acc.output.bytes + curr.output.bytes,
+									entries: acc.output.entries + curr.output.entries,
+								},
+							})),
+					);
+
+				return {
+					id: rawDetails.data.SearchInfo.ID,
+					userID: toNumericID(rawDetails.data.SearchInfo.UID),
+
+					filter,
+					finished: rawStats.data.Finished && rawDetails.data.Finished,
+
+					query: searchAttachMsg.data.Info.UserQuery,
+					effectiveQuery: searchAttachMsg.data.Info.EffectiveQuery,
+
+					metadata: searchAttachMsg.data.Info.Metadata ?? {},
+					entries: rawStats.data.EntryCount,
+					duration: rawDetails.data.SearchInfo.Duration,
+					start: new Date(rawDetails.data.SearchInfo.StartRange),
+					end: new Date(rawDetails.data.SearchInfo.EndRange),
+					minZoomWindow: rawDetails.data.SearchInfo.MinZoomWindow,
+					downloadFormats: rawDetails.data.SearchInfo.RenderDownloadFormats,
+					tags: searchAttachMsg.data.Info.Tags,
+
+					storeSize: rawDetails.data.SearchInfo.StoreSize,
+					processed: {
+						entries: pipeline[0]?.input?.entries ?? 0,
+						bytes: pipeline[0]?.input?.bytes ?? 0,
+					},
+
+					pipeline,
+				};
+			},
+		),
+	);
